Zepp OS: Refactor config, fix health on GTR 3 and GTS 3
The config refactor in addf7ff6a broke health settings on GTR3 and GTS3

- GTS 3 and GTR 3 health configs use protocol v1. The only difference
  seems to be that the steps goal is a SHORT instead of an INT.
- It needs a refactoring from the ground up to better handle different
  versions, but this is enough to get the GTR 3 and GTS 3 working.

+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/devices/huami/Huami2021Handler.java
@@ -17,5 +17,5 @@
 package nodomain.freeyourgadget.gadgetbridge.service.devices.huami;
 
 public interface Huami2021Handler {
-    void handle2021Payload(int type, byte[] payload);
+    void handle2021Payload(short type, byte[] payload);
 }
